requirements/common.txt
Flask==0.11.1
Jinja2==2.8
blinker==1.4
Werkzeug==0.11.10
itsdangerous==0.24
Flask-Bootstrap==3.3.6.0
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-Moment==0.5.1
Flask-WTF==0.12.0
WTForms==2.1
Mako==1.0.4
click==6.6
alembic==0.8.6
MarkupSafe==0.23
flask-login==0.3.2
flask-pagedown==0.2.1
markdown==2.6.6
bleach==1.4.3
six==1.10.0
Pillow==3.3.0

